instead of deleting tables one by one, drop the db and remake it in each test
