+++ TiFBackendUtils/SQLExecutable/index.ts
@@ -10,12 +10,13 @@ const EnvVarsSchema = z
   .object({
     DATABASE_HOST: z.string(),
     DATABASE_PASSWORD: z.string(),
-    DATABASE_USERNAME: z.string(),
-    DATABASE_NAME: z.string()
+    DATABASE_USERNAME: z.string()
   })
   .passthrough()
 
 const envVars = EnvVarsSchema.parse(process.env)
+
+const dbName = 'tif'
 
 async function createDatabaseConnection () {
   try {
@@ -23,7 +24,7 @@ async function createDatabaseConnection () {
       host: envVars.DATABASE_HOST,
       user: envVars.DATABASE_USERNAME,
       password: envVars.DATABASE_PASSWORD,
-      database: envVars.DATABASE_NAME,
+      database: dbName,
       timezone: "Z",
       namedPlaceholders: true,
       decimalNumbers: true
@@ -36,4 +37,24 @@ async function createDatabaseConnection () {
   }
 }
 
-export const conn = new SQLExecutable(createDatabaseConnection())
+const connectionPromise = createDatabaseConnection()
+
+export const conn = new SQLExecutable(connectionPromise)
+
+export const recreateDatabase = async () => {
+  const connection = await connectionPromise;
+
+  try {
+    await connection.changeUser({ database: undefined });
+
+    await connection.query(`DROP DATABASE ${dbName}`);
+
+    await connection.query(`CREATE DATABASE ${dbName}`);
+
+    await connection.changeUser({ database: dbName });
+
+    console.log('Reset the database successfully');
+  } catch (error) {
+    console.error('An error occurred:', error);
+  }
+};
